% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_detect_tissue_type.R
\name{auto_detect_tissue_type}
\alias{auto_detect_tissue_type}
\title{auto_detect_tissue_type: automatically detect a tissue type of the dataset}
\usage{
auto_detect_tissue_type(path_to_db_file, scRNAseqData, seurat_object, scaled)
}
\arguments{
\item{path_to_db_file}{xlsx: DB file with cell types}

\item{scRNAseqData}{Matrix: input scRNA-seq matrix (rownames - genes, column names - cells),}

\item{seurat_object}{Seurat Object: Seurat object from which the scRNAseqdata is derived from}

\item{scaled}{Boolean: indicates whether the matrix is scaled. TRUE by default.}
}
\description{
auto_detect_tissue_type: automatically detect a tissue type of the dataset
}
