% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc_rank_plot.R
\name{bc_rank_plot}
\alias{bc_rank_plot}
\title{Barcode rank / knee plot for filtering empty droplets}
\usage{
bc_rank_plot(stats, raw_cells, filt_cells, save)
}
\arguments{
\item{stats}{DataFrame: Output from `DropletUtil::barcodeRanks`}

\item{raw_cells}{Character vector: names of all unfiltered cells from raw_matrix}

\item{filt_cells}{Character vector: names of filtered cells}

\item{save}{String: Output directory}
}
\description{
Visualizes the inflection point to filter empty droplets.
}
