% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{filt_cells}
\alias{filt_cells}
\title{filtered_cells character vector}
\format{
Seurat object:
\describe{
  \item{test3}{test3}
  ...
}
}
\source{
\url{"https://google.com"}
}
\usage{
filt_cells
}
\description{
filtered_cells character vector
}
\keyword{datasets}
