% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{filt_matrix}
\alias{filt_matrix}
\title{Filtered count matrix}
\format{
Count matrix object:
\describe{
  \item{test2}{test2}
  ...
}
}
\source{
\url{"https://google.com"}
}
\usage{
filt_matrix
}
\description{
Filtered count matrix
}
\keyword{datasets}
