% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_sets_prepare.R
\name{gene_sets_prepare}
\alias{gene_sets_prepare}
\title{gene_sets_prepare: prepare gene sets and calculate marker sensitivity from input Cell Type excel file}
\usage{
gene_sets_prepare(path_to_db_file, cell_type)
}
\arguments{
\item{path_to_db_file}{xlsx: DB file with cell types}

\item{cell_type}{string: cell type (e.g. Immune system, Liver, Pancreas, Kidney, Eye, Brain)}
}
\description{
gene_sets_prepare: prepare gene sets and calculate marker sensitivity from input Cell Type excel file
}
