% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kb_stats}
\alias{kb_stats}
\title{kb_stats file}
\format{
JSON files from Kallisto:
\describe{
  \item{test3}{test3}
  ...
}
}
\source{
\url{"https://google.com"}
}
\usage{
kb_stats
}
\description{
kb_stats file
}
\keyword{datasets}
