% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc.1k.seurat}
\alias{pbmc.1k.seurat}
\title{Seurat object}
\format{
Seurat object:
\describe{
  \item{test3}{test3}
  ...
}
}
\source{
\url{"https://google.com"}
}
\usage{
pbmc.1k.seurat
}
\description{
Seurat object
}
\keyword{datasets}
