% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plain.R
\name{plain}
\alias{plain}
\title{Formatting to plain text}
\usage{
plain(x, ...)
}
\arguments{
\item{x}{string: text to be formatted}

\item{...}{further arguments passed to or from other methods}
}
\description{
Formatting to plain text
}
