% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pct_genes.R
\name{plot_pct_genes}
\alias{plot_pct_genes}
\title{Plot pct genes}
\usage{
plot_pct_genes(mat, tr2g, top_n = 20, symbol = "ensembl")
}
\arguments{
\item{mat}{Matrix/dGCmatrix: Counts matrix}

\item{tr2g}{Dataframe/tibble: Tr2g output from kb workflow, which maps ensembl transcript IDs to gene names and gene IDs.}

\item{top_n}{Integer: top n differentially expressed genes}

\item{symbol}{string: rownames/gene names. Can be either "ensembl" or "symbol".}
}
\description{
Plot genes with the highest proportions cells. Adapted from kallistobustools tutorial
}
