% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_html.R
\name{print_HTML}
\alias{print_HTML}
\title{Print and collate stats in html}
\usage{
print_HTML(
  seq_stats,
  cell_stats,
  dir,
  img_dir,
  sample_id,
  file_name = "_summary"
)
}
\arguments{
\item{seq_stats}{data.frame: sequencing statistics}

\item{cell_stats}{data.frame: cell statistics}

\item{dir}{string: Directory to save output .html file}

\item{img_dir}{string: Directory to find barcode rank plot for display}

\item{sample_id}{string: Name of sample under study. Can be NULL.}

\item{file_name}{string: Filename}
}
\description{
Gives a CellRanger-like html output.
}
