% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_cells}
\alias{raw_cells}
\title{raw_cells}
\format{
raw_cell character vector
\describe{
  \item{test3}{test3}
  ...
}
}
\source{
\url{"https://google.com"}
}
\usage{
raw_cells
}
\description{
raw_cells
}
\keyword{datasets}
