% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_matrix}
\alias{raw_matrix}
\title{Raw count matrix}
\format{
Count matrix object:
\describe{
  \item{test1}{test1}
  ...
}
}
\source{
\url{"https://google.com"}
}
\usage{
raw_matrix
}
\description{
Raw count matrix
}
\keyword{datasets}
