% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_count_output.R
\name{read_count_output}
\alias{read_count_output}
\title{Read Count Output}
\usage{
read_count_output(dir, name)
}
\arguments{
\item{dir}{string: path to folder containing files with `name` as prefix}

\item{name}{string: prefix of the gene/barcode/count files}
}
\value{
Labelled Count matrix object
}
\description{
Generate labelled count matrix
}
