% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sctype_score_.R
\name{sctype_score}
\alias{sctype_score}
\title{sctype_score: calculate ScType scores and assign cell types}
\usage{
sctype_score(
  scRNAseqData,
  scaled = TRUE,
  gs,
  gs2 = NULL,
  gene_names_to_uppercase = TRUE
)
}
\arguments{
\item{scRNAseqData}{matrix: input scRNA-seq matrix (rownames - genes, column names - cells),}

\item{scaled}{boolean: indicates whether the matrix is scaled (TRUE by default)}

\item{gs}{character vector: list of gene sets positively expressed in the cell type}

\item{gs2}{character vector: list of gene sets that should not be expressed in the cell type (NULL if not applicable)}

\item{gene_names_to_uppercase}{boolean: indicates whether to convert gene names to uppercase letters (TRUE by default)}
}
\description{
sctype_score: calculate ScType scores and assign cell types
}
